type change from uint64_t to uint32_t

+++ policies/dir_watcher.bpf.h
@@ -29,7 +29,7 @@ struct {
 
 struct {
     __uint(type, BPF_MAP_TYPE_HASH);
-    __type(key, __u64);
+    __type(key, __u32);
     __type(value, bool);
     __uint(max_entries, 128); // support up to 128 PIDs for now
 } pid_watchlist SEC(".maps");
@@ -44,7 +44,7 @@ static inline bool inode_in_watchlist(u64 inode_no) {
     return false;
 };
 
-static inline bool pid_in_watchlist(u64 pid) {
+static inline bool pid_in_watchlist(u32 pid) {
     u8 *ret = bpf_map_lookup_elem(&pid_watchlist, &pid);
     if (ret != NULL) {
         return true;
